Rename vector2.py to vector.py
